adding the click modifier
