+++ app/src/main/java/com/example/clicker/presentation/home/HomeViewModel.kt
@@ -60,7 +60,7 @@ class HomeViewModel(
             }
         }
     }
-    
+
 
     fun updateAuthenticationCode(token:String){
